refactor(effects): the doubles get their own file, so lib.rs stays under its ratchet
Binding the authority to its target pushed `portcullis-effects/src/lib.rs` from
under its 2813-line ceiling to 3002, and `check-line-ratchet.sh` refused it.
The ceiling only moves down, so the fix is to extract — not to raise it.

The seam is a real one rather than a convenient line. `RecordingEffects`,
`DenyAllEffects` and `AllowListEffects` perform no real I/O, which is exactly
the distinction `scripts/inert-authority-manifest.txt` already draws when it
classes those three scopes `N` and judges `RealEffects` by what it actually
does. lib.rs keeps the production effects and the policy wrapper; `doubles.rs`
takes the doubles. Public paths are unchanged — `RecordingEffects` and friends
are re-exported, so no caller moves.

`RecordingEffects` is the one that is not purely inert, and its module header
says so: its `run`/`commit`/`push`/`run_argv` and its `NetEffect::fetch` spend
the authority, because `PolicyEnforced` forwards those without spending.

Two things this turned up.

`cargo xtask inert-authority` keys rows on `(file, scope)`, so a scope that
moves file reads as a row with no sites left — and the message says "the debt
was paid", which is its only hypothesis for that shape and was wrong here. The
three rows are re-pointed at `doubles.rs` rather than deleted, and the manifest
header now names the third case so the next reader is not misled into lowering
`INERT_TOTAL` for a move.

The gate reads `git ls-files`, so an untracked new file is invisible to it: the
three rows still reported "no inert witness left" until `doubles.rs` was staged.
Worth knowing when a gate disagrees with the tree in front of you.

The manifest's own arithmetic was wrong and is corrected in passing: it said
"30 of the 40 sites are in one file" and the figure was 37. The per-scope
argument it was making is the right one and this split is the demonstration —
lib.rs now holds 6 and doubles.rs 25, and the RealEffects row going 6 -> 3 is a
statement about which mechanism changed in a way no file count could be.

`portcullis-effects/src/lib.rs` is now at or below its 2500-line target, which
the ratchet reports rather than merely tolerating.

Verified: `cargo fmt --check --all` clean; workspace clippy at `--all-features`
and `-p portcullis` at default both clean; `cargo test --workspace --exclude
nucleus-verifier-service --all-features` green, no failures; `check-line-ratchet.sh`,
`cargo xtask ci-spec check`, `inert-authority` (34 across 665 production files)
and `law-mechanisms` all ok.
