+++ scripts/inert-authority-manifest.txt
@@ -55,18 +55,25 @@
 #   CLASS | file | scope | count | note
 #
 # `scope` is the enclosing `impl` target (for `impl T for X`, that is `X`) or
-# `trait T`. Not the file: 31 of the 34 sites are in one file (3 + 12 + 7 + 6 +
-# 1 + 1 + 1, recomputed -- the earlier "30 of the 40" here did not add up
-# either, at 37) and they are not alike. DenyAllEffects refuses every call,
-# RecordingEffects writes a log line, and RealEffects::{run, commit, push} used
-# to spawn processes. A file-level allowance would have laundered the three
-# that mattered behind the twenty-eight that did not, which is exactly the
-# shape of laundering .dead-code-ratchet.toml refuses between crates.
+# `trait T`. Not the file, and portcullis-effects is why. Every one of these
+# rows but two used to name lib.rs -- 37 of the then-40 sites (the "30 of the
+# 40" written here originally did not add up) -- and they were not alike.
+# DenyAllEffects refuses every call, RecordingEffects writes a log line, and
+# RealEffects::{run, commit, push} spawned processes. A file-level allowance
+# would have laundered the three that mattered behind the thirty-four that did
+# not, which is exactly the shape of laundering .dead-code-ratchet.toml refuses
+# between crates.
 #
-# Those three are paid now, and the per-scope split is what made it possible to
+# Those three are paid now, and the per-scope key is what made it possible to
 # say so exactly: the RealEffects row went 6 -> 3, a statement about which
-# mechanism changed. A file count going 37 -> 31 would have said nothing at
-# all.
+# mechanism changed. A file count going 37 -> 31 would have said nothing.
+#
+# The key is (file, scope), so a scope that MOVES file reads to this gate as a
+# row with no sites left -- and its message says "the debt was paid", which is
+# its only hypothesis for that shape and was wrong when the doubles moved to
+# doubles.rs. Today lib.rs holds 6 (RealEffects 3, three trait declarations 1
+# each) and doubles.rs holds 25 (12 + 7 + 6). If a row ever reports paid,
+# confirm the code did not simply move before lowering INERT_TOTAL.
 #
 #   N -- the scope performs no act, so no authority is being discarded.
 #   D -- debt: the act happens and the witness is dropped.
@@ -109,9 +116,9 @@ WITNESS = [
 N | crates/portcullis-effects/src/lib.rs | RealEffects | 3 | WAS THE finding, at 6 and class D: run() spawned the parsed command, commit() ran `git add -u` and committed, push() ran `git push` -- each took Authority BY VALUE and dropped it, so the witness bounded nothing about which command, remote or branch. C2.4 paid it: all three now spend against the act they are about to perform (`Authority::spend_on`), and the mint side takes the target so the bundle can bind it. The three left are WebEffect::fetch/search and AgentSpawnEffect::spawn, whose bodies are Err(NotImplemented) -- nothing is performed, so there is no act for an authority to bound, and the group is no longer mixed.
 D | crates/nucleus-identity/src/ca/mod.rs | trait CaClient | 1 | sign_attested_csr's DEFAULT body is self.sign_csr(csr, private_key, identity, ttl) with `_attestation: &LaunchAttestation` unread, under a doc comment promising "The attestation is cryptographically bound to the certificate" and "Verifiers can require specific attestation hashes". SelfSignedCa overrides it (ca/self_signed.rs); SpireCaClient (ca/spire.rs) does not, so it inherits the version that binds nothing.
 D | crates/nucleus-ifc-kernel/src/ifc_api.rs | FlowTracker | 1 | reset_session_ceiling takes `_token: &SessionCleanseToken` by REFERENCE and never consumes it, then assigns the new ceiling. A cleanse token that survives its own use is not single-use; ADR 0006's C3.2 folds this into the one burn ledger.
-N | crates/portcullis-effects/src/lib.rs | DenyAllEffects | 12 | every method returns Err(PolicyDenied(..)) or Err(io::Error). Nothing is performed, so there is no act for an authority to bound. Dropping the witness here is correct, not deferred.
-N | crates/portcullis-effects/src/lib.rs | RecordingEffects | 7 | each method calls self.record(..) and returns; a test double that performs no effect. Same reasoning as DenyAllEffects. Was 10: run/commit/push now spend, matching RealEffects, because PolicyEnforced stopped spending on their behalf -- a double that dropped the authority where the real one spends would diverge on exactly the property the receipt tests check, and did, for one commit.
-N | crates/portcullis-effects/src/lib.rs | AllowListEffects | 6 | a simulation double: read() returns a canned file_read_response, write()/append() only check_path, glob()/search() return empty. No real I/O. Worth reading anyway as a miniature of the whole finding -- it enforces its OWN allow-list and never consults the authority it was handed.
+N | crates/portcullis-effects/src/doubles.rs | DenyAllEffects | 12 | every method returns Err(PolicyDenied(..)) or Err(io::Error). Nothing is performed, so there is no act for an authority to bound. Dropping the witness here is correct, not deferred.
+N | crates/portcullis-effects/src/doubles.rs | RecordingEffects | 7 | each method calls self.record(..) and returns; a test double that performs no effect. Same reasoning as DenyAllEffects. Was 10: run/commit/push now spend, matching RealEffects, because PolicyEnforced stopped spending on their behalf -- a double that dropped the authority where the real one spends would diverge on exactly the property the receipt tests check, and did, for one commit.
+N | crates/portcullis-effects/src/doubles.rs | AllowListEffects | 6 | a simulation double: read() returns a canned file_read_response, write()/append() only check_path, glob()/search() return empty. No real I/O. Worth reading anyway as a miniature of the whole finding -- it enforces its OWN allow-list and never consults the authority it was handed.
 N | crates/portcullis-effects/src/lib.rs | trait ShellEffect | 1 | run_argv's declaration, no body. The underscore is a naming choice in a signature; implementors bind it themselves and are counted at their own scope.
 N | crates/portcullis-effects/src/lib.rs | trait AsyncShellSpawnEffect | 1 | run_argv_async's declaration, no body. As above.
 N | crates/portcullis-effects/src/lib.rs | trait NetEffect | 1 | fetch's declaration, no body. As above.
